[proliferate] extract session creation domain helpers
