+++ desktop/src/hooks/sessions/session-creation-helpers.ts
@@ -1,16 +1,10 @@
 import type {
   ContentPart,
-  ModelRegistry,
-  ModelRegistryModel,
   PromptInputBlock,
-  Session,
-  WorkspaceSessionLaunchCatalog,
 } from "@anyharness/sdk";
 import type { ConnectorLaunchResolutionWarning } from "@/lib/domain/mcp/types";
 import { getLatencyFlowRequestHeaders } from "@/lib/infra/measurement/latency-flow";
 import { trackProductEvent } from "@/lib/integrations/telemetry/client";
-import { resolveCoworkDefaultSessionModeId } from "@/lib/domain/cowork/session-mode-defaults";
-import type { PausedSessionModelAvailability } from "@/hooks/sessions/workflows/use-session-model-availability-workflow";
 import type { MeasurementOperationId } from "@/lib/infra/measurement/debug-measurement";
 import type { PromptAttachmentSnapshot } from "@/lib/domain/chat/composer/prompt-attachment-snapshot";
 import { batchSessionStoreWrites } from "@/lib/infra/scheduling/react-batching";
@@ -20,24 +14,6 @@ import {
 } from "@/stores/sessions/session-records";
 import { useSessionSelectionStore } from "@/stores/sessions/session-selection-store";
 import type { SessionRuntimeRecord } from "@/stores/sessions/session-types";
-
-export function resolveSessionCreationModeId(input: {
-  explicitModeId?: string | null;
-  workspaceSurface: string | null | undefined;
-  agentKind: string;
-  preferredModeId?: string | null;
-}): string | undefined {
-  const explicitModeId = input.explicitModeId?.trim() || undefined;
-  if (explicitModeId) {
-    return explicitModeId;
-  }
-
-  if (input.workspaceSurface === "cowork") {
-    return resolveCoworkDefaultSessionModeId(input.agentKind);
-  }
-
-  return input.preferredModeId?.trim() || undefined;
-}
 
 export function buildLatencyRequestOptions(latencyFlowId?: string | null) {
   const headers = getLatencyFlowRequestHeaders(latencyFlowId);
@@ -150,97 +126,4 @@ export function reportConnectorLaunchWarnings(
   }
 
   showToast(`${warnings.length} connectors weren't available in this session.`, "info");
-}
-
-function modelMatchesId(model: Pick<ModelRegistryModel, "id" | "aliases">, modelId: string): boolean {
-  return model.id === modelId || (model.aliases ?? []).includes(modelId);
-}
-
-function findRegistryModel(
-  registries: ModelRegistry[],
-  agentKind: string,
-  modelId: string,
-) {
-  const registry = registries.find((candidate) => candidate.kind === agentKind) ?? null;
-  const model = registry?.models.find((candidate) => modelMatchesId(candidate, modelId)) ?? null;
-  return { registry, model };
-}
-
-function launchCatalogExposesModel(input: {
-  launchCatalog: WorkspaceSessionLaunchCatalog;
-  agentKind: string;
-  requestedModelId: string;
-  requestedModel: ModelRegistryModel;
-}): boolean {
-  const agent = input.launchCatalog.agents.find((candidate) =>
-    candidate.kind === input.agentKind
-  );
-  if (!agent) {
-    return false;
-  }
-
-  const acceptedIds = new Set([
-    input.requestedModelId,
-    input.requestedModel.id,
-    ...(input.requestedModel.aliases ?? []),
-  ]);
-  return agent.models.some((model) => acceptedIds.has(model.id));
-}
-
-export function hasImmediateLaunchModelMismatch(input: {
-  session: Session;
-  agentKind: string;
-  registries: ModelRegistry[];
-  launchCatalog: WorkspaceSessionLaunchCatalog;
-}): boolean {
-  const requestedModelId = input.session.requestedModelId;
-  const currentModelId = input.session.modelId;
-  if (!requestedModelId || !currentModelId || requestedModelId === currentModelId) {
-    return false;
-  }
-
-  const requested = findRegistryModel(
-    input.registries,
-    input.agentKind,
-    requestedModelId,
-  );
-  if (!requested.model?.launchRemediation) {
-    return false;
-  }
-
-  return !launchCatalogExposesModel({
-    launchCatalog: input.launchCatalog,
-    agentKind: input.agentKind,
-    requestedModelId,
-    requestedModel: requested.model,
-  });
-}
-
-export function buildPausedModelAvailability(input: {
-  session: Session;
-  workspaceId: string;
-  agentKind: string;
-  registries: ModelRegistry[];
-}): PausedSessionModelAvailability {
-  const requestedModelId = input.session.requestedModelId ?? "";
-  const currentModelId = input.session.modelId ?? "";
-  const requested = findRegistryModel(input.registries, input.agentKind, requestedModelId);
-  const current = findRegistryModel(input.registries, input.agentKind, currentModelId);
-  const providerDisplayName =
-    requested.registry?.displayName
-    ?? current.registry?.displayName
-    ?? input.agentKind;
-
-  return {
-    id: `${input.session.id}:${requestedModelId}:${currentModelId}`,
-    sessionId: input.session.id,
-    workspaceId: input.workspaceId,
-    agentKind: input.agentKind,
-    providerDisplayName,
-    requestedModelId,
-    requestedModelDisplayName: requested.model?.displayName ?? requestedModelId,
-    currentModelId,
-    currentModelDisplayName: current.model?.displayName ?? currentModelId,
-    remediation: requested.model?.launchRemediation ?? null,
-  };
 }
